Update and rename repro_test.go to config_test.go
